backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.5.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-generativeai==0.3.0
gtts==2.4.0
SpeechRecognition==3.10.0
spacy==3.7.2
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
cors==1.0.1
pyaudio==0.2.13
motor==3.3.2
pymongo[srv]==4.5.0
dnspython==2.4.2
certifi==2023.11.17